trunk=0.8.2 (rd235 experimental branch merge, one year of developments)

+++ xmview-os/xmview/um_select.c
@@ -288,7 +288,7 @@ int wrap_out_select(int sc_number,struct pcb *pc)
 	if (sd != NULL) {
 		epoch_t oldepoch=um_setnestepoch(0);
 		register int n=pc->sysargs[0];
-		int pfds[3];
+		long pfds[3];
 		fd_set lfds[3]; /* local copy of the signaled SC fds */
 		int i,j,fd;
 		pc->retval=getrv(pc);
